requirements.txt
dateparser>=0.7.0
python-telegram-bot[socks]>=11.1
